Move more types into worker-interface where they are used in giant
