+++ packages/common/src/types.ts
@@ -1,6 +1,16 @@
 import { z } from 'zod';
 import { outputLanguageCodes } from './languages';
-import { InputLanguageCode, LlmBackend } from './worker-interface-types';
+import {
+	InputLanguageCode,
+	LlmBackend,
+	LLMOutputFailure,
+	LLMOutputSuccess,
+	LlmPrompt,
+	OutputBase,
+	TranscriptionOutputSuccess,
+	TranscriptionOutputFailure,
+	TranscriptionResult,
+} from './worker-interface-types';
 
 export const OutputLanguageCode = z.enum(outputLanguageCodes);
 export type OutputLanguageCode = z.infer<typeof OutputLanguageCode>;
@@ -87,37 +97,6 @@ export const WebpageSnapshot = z.object({
 });
 export type WebpageSnapshot = z.infer<typeof WebpageSnapshot>;
 
-const OutputBase = z.object({
-	id: z.string(),
-	userEmail: z.string(),
-});
-
-const TranscriptionOutputBase = OutputBase.extend({
-	originalFilename: z.string(),
-});
-
-export const TranscriptionOutputSuccess = TranscriptionOutputBase.extend({
-	// status must be kept in sync with https://github.com/guardian/giant/blob/main/backend/app/extraction/ExternalTranscriptionExtractor.scala#L76
-	status: z.literal('SUCCESS'),
-	languageCode: OutputLanguageCode,
-	combinedOutputKey: z.string(),
-	duration: z.optional(z.number()),
-	maybeEnqueuedAtEpochMillis: z.optional(z.number()),
-	includesTranslation: z.boolean(),
-	translationRequested: z.boolean(),
-});
-
-export const LLMOutputSuccess = OutputBase.extend({
-	status: z.literal('LLM_SUCCESS'),
-	outputKey: z.string(),
-});
-export type LLMOutputSuccess = z.infer<typeof LLMOutputSuccess>;
-
-export const LLMOutputFailure = OutputBase.extend({
-	status: z.literal('LLM_FAILURE'),
-});
-export type LLMOutputFailure = z.infer<typeof LLMOutputFailure>;
-
 export const MediaDownloadFailureReason = z.union([
 	z.literal('FAILURE'),
 	z.literal('INVALID_URL'),
@@ -135,12 +114,6 @@ export const MediaDownloadFailure = OutputBase.extend({
 
 export type MediaDownloadFailure = z.infer<typeof MediaDownloadFailure>;
 
-export const TranscriptionOutputFailure = TranscriptionOutputBase.extend({
-	// status must be kept in sync with https://github.com/guardian/giant/blob/main/backend/app/extraction/ExternalTranscriptionExtractor.scala#L76
-	status: z.literal('TRANSCRIPTION_FAILURE'),
-	noAudioDetected: z.boolean(),
-});
-
 export const TranscriptionOutput = z.union([
 	TranscriptionOutputSuccess,
 	TranscriptionOutputFailure,
@@ -148,14 +121,6 @@ export const TranscriptionOutput = z.union([
 	LLMOutputSuccess,
 	LLMOutputFailure,
 ]);
-
-export type TranscriptionOutputSuccess = z.infer<
-	typeof TranscriptionOutputSuccess
->;
-
-export type TranscriptionOutputFailure = z.infer<
-	typeof TranscriptionOutputFailure
->;
 
 export const transcriptionOutputIsSuccess = (
 	output: TranscriptionOutput,
@@ -311,14 +276,6 @@ export type TranscribeFileRequestBody = z.infer<
 	typeof transcribeFileRequestBody
 >;
 
-export const LlmPrompt = z.object({
-	system: z.string().optional(),
-	user: z.string(),
-	assistant: z.string().optional(),
-});
-
-export type LlmPrompt = z.infer<typeof LlmPrompt>;
-
 export const LlmRequestBody = z.object({
 	prompt: LlmPrompt,
 	backend: LlmBackend,
@@ -399,13 +356,6 @@ export const TranscriptionMetadata = z.object({
 	totalTimeMs: z.optional(z.number()),
 });
 export type TranscriptionMetadata = z.infer<typeof TranscriptionMetadata>;
-
-export const TranscriptionResult = z.object({
-	transcripts: Transcripts,
-	transcriptTranslations: z.optional(Transcripts),
-	metadata: TranscriptionMetadata,
-});
-export type TranscriptionResult = z.infer<typeof TranscriptionResult>;
 
 export const TranscriptionItemWithTranscript = z.object({
 	item: TranscriptionDynamoItem,
